+++ app/(pages)/movie/[id]/page.tsx
@@ -7,7 +7,7 @@ import {
   DetailRelatedSection,
 } from "@/app/features/movie/sections";
 import { DetailReviewSection } from "@/app/features/review/sections";
-import { getMovieDetail, saveMovie } from "@/app/features/movie/apis";
+import { getMovieDetail, saveMovie } from "@/app/features/movie/apis/post";
 
 export default async function Page({
   params,
