--- app/features/auth/components/index.ts
@@ -1,1 +1,0 @@
-export * from "./login-modal";
