+++ app/features/auth/login/components/index.ts
@@ -1,0 +1,1 @@
+export { LoginModal } from "./login-modal";
