+++ app/features/auth/login/models/index.ts
@@ -1,0 +1,1 @@
+export type { User } from "./auth";
