+++ app/features/like/apis/index.ts
@@ -1,0 +1,2 @@
+export { toggleLikeMovie } from "./toggle-like-movie";
+export { toggleLikeReview } from "./toggle-like-review";
