+++ app/features/like/components/index.ts
@@ -1,0 +1,1 @@
+export { LikeButton } from "./like-button";
