+++ app/features/movie/apis/index.ts
@@ -1,8 +1,7 @@
-export * from "./get-movie-detail-info";
-export * from "./get-movie-trailer";
-export * from "./get-movies-by-genre";
-export * from "./get-movies-by-now-playing";
-export * from "./get-movies-by-top-rated";
-export * from "./get-movies-by-upcoming";
-export * from "./save-movie-info";
-export * from "./toggle-like-movie";
+export { getMovieDetail } from "./get-movie-detail-info";
+export { getMovieTrailer } from "./get-movie-trailer";
+export { getMoviesByGenre } from "./get-movies-by-genre";
+export { getMoviesByNowPlaying } from "./get-movies-by-now-playing";
+export { getMoviesByTopRated } from "./get-movies-by-top-rated";
+export { getMoviesByUpcoming } from "./get-movies-by-upcoming";
+export { saveMovie } from "./save-movie-info";
