+++ app/features/movie/components/index.ts
@@ -1,5 +1,5 @@
-export * from "./movie-card";
-export * from "./main-banner-carousel";
-export * from "./main-carousel";
-export * from "./credits-card";
-export * from "./trailer-iframe";
+export { MovieCard } from "./movie-card";
+export { MainBannerCarousel } from "./main-banner-carousel";
+export { MainCarousel } from "./main-carousel";
+export { CreditsCard } from "./credits-card";
+export { TrailerIframe } from "./trailer-iframe";
