+++ app/features/review/apis/index.ts
@@ -1,9 +1,8 @@
-export * from "./deleteReview";
-export * from "./toggle-like-review";
-export * from "./post-new-review";
-export * from "./update-review";
-export * from "./get-my-review-by-movie";
-export * from "./get-all-reviews-by-movie";
-export * from "./get-recent-reviews-to-main";
-export * from "./get-reviews-by-movie-limit";
-export * from "./get-my-reviews-to-mypage";
+export { postNewReview } from "./post-new-review";
+export { updateReview } from "./update-review";
+export { deleteReview } from "./deleteReview";
+export { getMyReviewByMovie } from "./get-my-review-by-movie";
+export { getReviewsByMovie } from "./get-all-reviews-by-movie";
+export { getRecentReviews } from "./get-recent-reviews-to-main";
+export { getReviewsByMovieLimit } from "./get-reviews-by-movie-limit";
+export { getMyReviewsToMyPage } from "./get-my-reviews-to-mypage";
