+++ app/features/review/components/index.ts
@@ -1,10 +1,11 @@
-export * from "./main-review-card";
-export * from "./main-review-carousel";
-export * from "./review-form";
-export * from "./review-action-buttons";
-export * from "./review-delete-dialog";
-export * from "./detail-review-card";
-export * from "./detail-review-list";
-export * from "./detail-review-dialog";
-export * from "./more-review-button";
-export * from "./review-card-skeleton";
+export { MainReviewCard } from "./main-review-card";
+export { MainReviewCarousel } from "./main-review-carousel";
+export { ReviewForm } from "./review-form";
+export { ReviewActionButtons } from "./review-action-buttons";
+export { ReviewDeleteDialog } from "./review-delete-dialog";
+export { DetailReviewCard } from "./detail-review-card";
+export { DetailReviewList } from "./detail-review-list";
+export { DetailReviewDialog } from "./detail-review-dialog";
+export { MoreReviewButton } from "./more-review-button";
+export { ReviewCardSkeleton } from "./review-card-skeleton";
+export { StarRating } from "./star-rating";
