+++ app/features/user/apis/index.ts
@@ -1,3 +1,2 @@
-export * from "./getMyActives";
-export * from "./getMyWishMovies";
-export * from "./withraw";
+export { getMyActives } from "./get-my-actives";
+export { getMyWishMovies } from "./get-my-wish-movies";
