+++ app/shared/actions/index.ts
@@ -1,2 +1,1 @@
-export * from "./cookie";
 export * from "./redirect";
