+++ app/shared/components/index.ts
@@ -1,7 +1,5 @@
 export * from "./ui";
 export * from "./layout";
 export * from "./blocks";
-export * from "./rating";
 export * from "./fallback";
-export * from "./like";
 export * from "./async";
