--- app/shared/components/like/index.ts
@@ -1,1 +1,0 @@
-export * from "./like-button";
