--- app/shared/components/rating/index.ts
@@ -1,1 +1,0 @@
-export * from "./star-rating";
