--- app/shared/constants/token.ts
@@ -1,4 +1,0 @@
-export const TOKEN = {
-  accessToken: "accessToken",
-  refreshToken: "refreshToken",
-};
